+++ packages/algoliasearch-helper/test/run.js
@@ -38,11 +38,16 @@ jest.runCLI(dynamicJestConfig, projectsRootPaths).then(function (response) {
       process.env.INTEGRATION_TEST_APPID,
       process.env.INTEGRATION_TEST_API_KEY
     );
-    client.deleteIndex =
-      client.deleteIndex ||
-      function (deleteIndexName) {
+    var originalDeleteIndex = client.deleteIndex.bind(client);
+    client.deleteIndex = function (deleteIndexName) {
+      if (!client.deleteIndex) {
         return client.initIndex(deleteIndexName).delete();
-      };
+      }
+      if (!client.initIndex) {
+        return originalDeleteIndex({ indexName: deleteIndexName });
+      }
+      return originalDeleteIndex(deleteIndexName);
+    };
     client.listIndexes = client.listIndexes || client.listIndices;
 
     client.listIndexes().then((content) => {
